+++ app/Http/Controllers/Admin/Team/TeamHeadController.php
@@ -1,22 +1,20 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Admin\Team;
 
 use App\Http\Controllers\Controller;
-use App\Models\HomeSlider;
 use App\Models\Publish;
+use App\Models\TeamSlider;
 use Illuminate\Http\Request;
 use App\Traits\GetLangMessage;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Validator;
-// use Intervention\Image\ImageManager;
-// use Intervention\Image\Drivers\Gd\Driver;
 use Illuminate\Support\Facades\File;
 use Exception;
 use Illuminate\Support\Facades\DB;
-use App\Traits\GetBoolFromDb;
+use App\Traits\GetBoolFromDB;
 
-class HomeSliderController extends Controller
+class TeamHeadController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -26,29 +24,27 @@ class HomeSliderController extends Controller
     public function index()
     {
         try {
-            $src = DB::table('home_sliders')->orderBy('ranking')->get();
+            $src = DB::table('team_sliders')->orderBy('ranking')->get();
             $publish = DB::table('publishes')->get();
 
-            $i = 1;
             $slideIds = [];
 
             foreach ($src as $slider) {
                 $slideIds[$slider->ranking] = $slider->id;
             }
 
-            return view('admin.home.index', [
-                'i' => $i,
+            return view('admin.team.head.index', [
                 'src' => $src,
-                'public' => GetBoolFromDb::getBool($publish, 'home.slider'),
+                'public' => GetBoolFromDB::getBool($publish, 'team.slider'),
                 'slideIds' => $slideIds,
             ]);
         } catch (Exception $e) {
 
-            return view('admin.home.index', compact('e'));
+            return view('admin.team.head.index', compact('e'));
         }
         $err = GetLangMessage::languagePackage('en')->databaseError;
 
-        return view('admin.home.index', compact('err'));
+        return view('admin.team.head.index', compact('err'));
     }
 
     /**
@@ -61,7 +57,7 @@ class HomeSliderController extends Controller
     {
         try {
             $credentials = Validator::make($request->all(), [
-                'title' => 'required|min:3|max:255|unique:home_sliders',
+                'title' => 'required|min:3|max:255|unique:team_sliders',
                 'image' => 'required|mimes:jpg,png,webp,jpeg',
             ]);
 
@@ -77,17 +73,12 @@ class HomeSliderController extends Controller
             $extension = strtolower($file->getClientOriginalExtension());
             $img_name = $name . '.' . $extension;
 
-            $upload_location = 'uploads/home_slider/';
+            $upload_location = 'uploads/team_slider/';
             $save_url = $upload_location . $img_name;
 
             $file->move($upload_location, $img_name);
 
-            // $manager = new ImageManager(new Driver());
-            // $read_img = $manager->read($save_url);
-            // $resize_img = $read_img->resize(300, 200);
-            // $resize_img->toJpeg(85)->save(base_path($save_url));
-
-            HomeSlider::insert([
+            TeamSlider::insert([
                 'title' => mb_strtolower(str_replace(' ', '-', $request->title)),
                 'image' => $save_url,
                 'ranking' => $request->ranking,
@@ -95,19 +86,49 @@ class HomeSliderController extends Controller
             ]);
 
             return redirect()->back()->with([
+                'present' => true,
                 'status' => true,
                 'message' => GetLangMessage::languagePackage('en')->uploadTrue,
             ]);
         } catch (Exception $e) {
             return redirect()->back()->with([
-                'status' => true,
+                'present' => true,
+                'status' => false,
                 'message' => GetLangMessage::languagePackage('en')->uploadFalse,
             ]);
         }
 
         return redirect()->back()->with([
-            'status' => true,
+            'present' => true,
+            'status' => false,
             'message' => GetLangMessage::languagePackage('en')->uploadFalse,
+        ]);
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
+    {
+        try {
+            $slide = DB::table('team_sliders')->find($id);
+
+            return view('admin.team.head.edit_slide', compact('slide'));
+        } catch (Exception $e) {
+            return redirect()->back()->with([
+                'present' => true,
+                'status' => false,
+                'message' => GetLangMessage::languagePackage('en')->deleteFalse,
+            ]);
+        }
+
+        return redirect()->back()->with([
+            'present' => true,
+            'status' => false,
+            'message' => GetLangMessage::languagePackage('en')->deleteFalse,
         ]);
     }
 
@@ -138,7 +159,7 @@ class HomeSliderController extends Controller
             $extension = strtolower($file->getClientOriginalExtension());
             $img_name = $name . '.' . $extension;
 
-            $upload_location = 'uploads/home_slider/';
+            $upload_location = 'uploads/team_slider/';
             $save_url = $upload_location . $img_name;
 
             if (File::exists($request->old_image)) {
@@ -147,25 +168,28 @@ class HomeSliderController extends Controller
 
             $file->move($upload_location, $img_name);
 
-            HomeSlider::whereId($id)->update([
+            TeamSlider::whereId($id)->update([
                 'title' => mb_strtolower(str_replace(' ', '-', $request->title)),
                 'image' => $save_url,
                 'updated_at' => Carbon::now(),
             ]);
 
-            return redirect('dashboard')->with([
+            return redirect('admin.team.index')->with([
+                'present' => true,
                 'status' => true,
                 'message' => GetLangMessage::languagePackage('en')->updateTrue,
             ]);
         } catch (Exception $e) {
             return redirect()->back()->with([
-                'status' => true,
+                'present' => true,
+                'status' => false,
                 'message' => GetLangMessage::languagePackage('en')->updateFalse,
             ]);
         }
 
         return redirect()->back()->with([
-            'status' => true,
+            'present' => true,
+            'status' => false,
             'message' => GetLangMessage::languagePackage('en')->updateFalse,
         ]);
     }
@@ -183,14 +207,13 @@ class HomeSliderController extends Controller
         ]);
 
         if ($credentials->fails()) {
-            return redirect()->back()->with([
-                'status' => true,
-                'message' => GetLangMessage::languagePackage('en')->updateFalse,
-            ]);
+            return redirect()->back()
+                ->withErrors($credentials->errors())
+                ->withInput();
         }
 
         try {
-            Publish::where('name', 'home.slider')->update([
+            Publish::where('name', 'team.slider')->update([
                 'public' => $request->slideshow,
                 'updated_at' => Carbon::now(),
             ]);
@@ -198,13 +221,15 @@ class HomeSliderController extends Controller
             return redirect()->back();
         } catch (Exception $e) {
             return redirect()->back()->with([
-                'status' => true,
+                'present' => true,
+                'status' => false,
                 'message' => $e,
             ]);
         }
 
         return redirect()->back()->with([
-            'status' => true,
+            'present' => true,
+            'status' => false,
             'message' => GetLangMessage::languagePackage('en')->updateFalse,
         ]);
     }
@@ -219,19 +244,20 @@ class HomeSliderController extends Controller
     public function up(Request $request, $id)
     {
         try {
-            HomeSlider::whereId($request->previous_id)->update([
+            TeamSlider::whereId($request->previous_id)->update([
                 'ranking' => $request->ranking + 1,
                 'updated_at' => Carbon::now(),
             ]);
         } catch (Exception $e) {
             return redirect()->back()->with([
-                'status' => true,
+                'present' => true,
+                'status' => false,
                 'message' => $e,
             ]);
         }
 
         try {
-            HomeSlider::whereId($id)->update([
+            TeamSlider::whereId($id)->update([
                 'ranking' => intval($request->ranking),
                 'updated_at' => Carbon::now(),
             ]);
@@ -239,13 +265,15 @@ class HomeSliderController extends Controller
             return redirect()->back();
         } catch (Exception $e) {
             return redirect()->back()->with([
-                'status' => true,
+                'present' => true,
+                'status' => false,
                 'message' => $e,
             ]);
         }
 
         return redirect()->back()->with([
-            'status' => true,
+            'present' => true,
+            'status' => false,
             'message' => GetLangMessage::languagePackage('en')->updateFalse,
         ]);
     }
@@ -260,19 +288,20 @@ class HomeSliderController extends Controller
     public function down(Request $request, $id)
     {
         try {
-            HomeSlider::whereId($request->next_id)->update([
+            TeamSlider::whereId($request->next_id)->update([
                 'ranking' => $request->ranking - 1,
                 'updated_at' => Carbon::now(),
             ]);
         } catch (Exception $e) {
             return redirect()->back()->with([
-                'status' => true,
+                'present' => true,
+                'status' => false,
                 'message' => $e,
             ]);
         }
 
         try {
-            HomeSlider::whereId($id)->update([
+            TeamSlider::whereId($id)->update([
                 'ranking' => intval($request->ranking),
                 'updated_at' => Carbon::now(),
             ]);
@@ -280,13 +309,15 @@ class HomeSliderController extends Controller
             return redirect()->back();
         } catch (Exception $e) {
             return redirect()->back()->with([
-                'status' => true,
+                'present' => true,
+                'status' => false,
                 'message' => $e,
             ]);
         }
 
         return redirect()->back()->with([
-            'status' => true,
+            'present' => true,
+            'status' => false,
             'message' => GetLangMessage::languagePackage('en')->updateFalse,
         ]);
     }
@@ -300,7 +331,17 @@ class HomeSliderController extends Controller
     public function destroy($id)
     {
         try {
-            $slide = HomeSlider::findOrFail($id);
+            $check = TeamSlider::all();
+
+            if (count($check) === 1) {
+                return redirect()->back()->with([
+                    'present' => true,
+                    'status' => false,
+                    'message' => GetLangMessage::languagePackage('en')->methodError . ' At least 1 photo must be available.',
+                ]);
+            }
+
+            $slide = TeamSlider::findOrFail($id);
             $slide->delete();
 
             if (File::exists($slide->image)) {
@@ -308,18 +349,21 @@ class HomeSliderController extends Controller
             }
 
             return redirect()->back()->with([
+                'present' => true,
                 'status' => true,
                 'message' => GetLangMessage::languagePackage('en')->deleteTrue,
             ]);
         } catch (Exception $e) {
             return redirect()->back()->with([
-                'status' => true,
+                'present' => true,
+                'status' => false,
                 'message' => GetLangMessage::languagePackage('en')->deleteFalse,
             ]);
         }
 
         return redirect()->back()->with([
-            'status' => true,
+            'present' => true,
+            'status' => false,
             'message' => GetLangMessage::languagePackage('en')->deleteFalse,
         ]);
     }
